@@ -1,4 +1,4 @@
-// swift-tools-version:5.7
+// swift-tools-version:5.8
 
 import PackageDescription
 
@@ -9,9 +9,14 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0"),
+        .package(url: "https://github.com/apple/swift-http-types.git", from: "1.0.0"),
     ],
     targets: [
-        .target(name: "AWSLambdaEvents", dependencies: []),
-        .testTarget(name: "AWSLambdaEventsTests", dependencies: ["AWSLambdaEvents"]),
+        .target(name: "AWSLambdaEvents",
+                dependencies: [.product(name: "HTTPTypes", package: "swift-http-types")],
+                swiftSettings: [.enableExperimentalFeature("StrictConcurrency=complete")]),
+        .testTarget(name: "AWSLambdaEventsTests",
+                    dependencies: ["AWSLambdaEvents"],
+                    swiftSettings: [.enableExperimentalFeature("StrictConcurrency=complete")]),
     ]
 )
