@@ -16,8 +16,8 @@ import struct Foundation.Date
 
 // https://docs.aws.amazon.com/lambda/latest/dg/with-sns.html
 
-public struct SNSEvent: Decodable {
-    public struct Record: Decodable {
+public struct SNSEvent: Decodable, Sendable {
+    public struct Record: Decodable, Sendable {
         public let eventVersion: String
         public let eventSubscriptionArn: String
         public let eventSource: String
@@ -37,8 +37,8 @@ public struct SNSEvent: Decodable {
         case records = "Records"
     }
 
-    public struct Message {
-        public enum Attribute {
+    public struct Message: Sendable {
+        public enum Attribute: Sendable {
             case string(String)
             case binary([UInt8])
         }
@@ -68,9 +68,9 @@ extension SNSEvent.Message: Decodable {
         case messageAttributes = "MessageAttributes"
         case signatureVersion = "SignatureVersion"
         case timestamp = "Timestamp"
-        case signingCertURL = "SigningCertURL"
+        case signingCertURL = "SigningCertUrl"
         case message = "Message"
-        case unsubscribeURL = "UnsubscribeURL"
+        case unsubscribeURL = "UnsubscribeUrl"
         case subject = "Subject"
     }
 }
@@ -104,10 +104,3 @@ extension SNSEvent.Message.Attribute: Decodable {
         }
     }
 }
-
-#if swift(>=5.6)
-extension SNSEvent: Sendable {}
-extension SNSEvent.Record: Sendable {}
-extension SNSEvent.Message: Sendable {}
-extension SNSEvent.Message.Attribute: Sendable {}
-#endif
