@@ -14,16 +14,16 @@
 
 // https://docs.aws.amazon.com/lambda/latest/dg/with-sqs.html
 
-public struct SQSEvent: Decodable {
+public struct SQSEvent: Decodable, Sendable {
     public let records: [Message]
 
     enum CodingKeys: String, CodingKey {
         case records = "Records"
     }
 
-    public struct Message {
+    public struct Message: Sendable {
         /// https://docs.aws.amazon.com/AWSSimpleQueueService/latest/APIReference/API_MessageAttributeValue.html
-        public enum Attribute {
+        public enum Attribute: Sendable {
             case string(String)
             case binary([UInt8])
             case number(String)
@@ -93,9 +93,3 @@ extension SQSEvent.Message.Attribute: Decodable {
         }
     }
 }
-
-#if swift(>=5.6)
-extension SQSEvent: Sendable {}
-extension SQSEvent.Message: Sendable {}
-extension SQSEvent.Message.Attribute: Sendable {}
-#endif
